requirements.txt
Django==3.2.6
Pillow==8.3.2
pytz==2022.7.1
sqlparse==0.4.3
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
requests==2.28.2
urllib3==1.26.14
SecretStorage==3.3.3
pycryptodome==3.19.0